FU-261 fix: fix UI
